Corrected most of -Wall's complaints.

+++ src/blorb/blorblib.c
@@ -226,7 +226,7 @@ static bb_err_t bb_initialize_map(bb_map_t *map)
                     bb_resdesc_t *resources;
                     bb_resdesc_t **ressorted;
 
-                    if (len != numres*12+4)
+                    if (len != (unsigned int) numres*12+4)
                         return bb_err_Format; /* bad length field */
 
                     resources = (bb_resdesc_t *)malloc(numres * sizeof(bb_resdesc_t));
@@ -362,7 +362,7 @@ static bb_err_t bb_initialize_map(bb_map_t *map)
                 if (len) {
                     bb_aux_pict_t *aux = (bb_aux_pict_t *)malloc(len * sizeof(bb_aux_pict_t));
 
-                    for (jx=0; jx<len; jx++, ptr += 7) {
+                    for (jx=0; (unsigned int) jx<len; jx++, ptr += 7) {
                         bb_result_t res;
 
                         err = bb_load_resource(map, bb_method_DontLoad, &res,
@@ -403,7 +403,7 @@ static bb_err_t bb_initialize_map(bb_map_t *map)
                 if (len) {
                     bb_aux_sound_t *aux = (bb_aux_sound_t *)malloc(len * sizeof(bb_aux_sound_t));
 
-                    for (jx=0; jx<len; jx++, ptr += 2) {
+                    for (jx=0; (unsigned int) jx<len; jx++, ptr += 2) {
                         bb_result_t res;
 
                         err = bb_load_resource(map, bb_method_DontLoad, &res,
